% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{add.alpha}
\alias{add.alpha}
\title{Add transparency to any color}
\usage{
add.alpha(col, alpha = 0.5)
}
\arguments{
\item{col}{Color value or vector of colors}

\item{alpha}{Opacity value to apply to the color(s)}
}
\value{
A character vector containing color hex codes.
}
\description{
Add transparency to any color
}
\examples{
add.alpha("red",0.8)
}
