% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages_archosauria}
\alias{ages_archosauria}
\title{ages_archosauria}
\format{
A matrix with 13 rows and 2 collumns containing:
\describe{
\item{FAD}{Earliest occurrence age}
\item{LAD}{Latest occurrence age}
}
…for each taxon
}
\usage{
ages_archosauria
}
\description{
A dataset containing earliest and latest occurrence dates for clades shown in the example phylogeny.
}
\keyword{datasets}
