% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{archosauria}
\alias{archosauria}
\title{archosauria}
\format{
A list() object containing 15 species tables (data.frames) with the following data in each:
\describe{
\item{tna}{taxon names (species names)}
\item{max}{maximum ages}
\item{min}{minimum ages}
\item{ma}{mean ages}
}
}
\source{
Generated from data downloaded from the paleobiology database \url{https://paleobiodb.org} using the functions pdb(), occ.cleanup() and mk.sptab()
}
\usage{
archosauria
}
\description{
A dataset of stratigraphic ranges of species within the clades in tree_archosauria.
}
\keyword{datasets}
