% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{darken}
\alias{darken}
\title{Darken or lighten colors by adding/subtracting to or hsv channel values}
\usage{
darken(x, add = 0, abs = NULL)
}
\arguments{
\item{x}{Color value or vector of colors}

\item{add}{Value to be added to the third hsv-channel. Can be a vector of length x, or a vector of any length if length(x)==1}

\item{abs}{Value to substitute for the third hsv-channel. If set, this overrides the setting for parameter add. Can be a vector of length x, or a vector of any length if length(x)==1}
}
\value{
A color value or vector of color values of length x (or, if length(x)==1, the length of add or abs)
}
\description{
Darken or lighten colors by adding/subtracting to or hsv channel values
}
\examples{
darken(ggcol(3),abs=0.5)
}
