% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diversity_table}
\alias{diversity_table}
\title{diversity_table}
\format{
A data.frame() containing mean ages and diversity figures by stage.
\describe{
\item{x_orig}{ages for each stage in the phanerozoic}
\item{x}{ages converted for plotting on tree_archosauria, using the tsconv()-function}
\item{Sauroporomorpha}{diversity by stage for Sauropodomorpha}
\item{etc}{diversity by stage for each of the taxa represented in tree_archosauria}
...
}
}
\usage{
diversity_table
}
\description{
A dataset of diversity by stage, exemplifying the output produced by the divDyn-package.
}
\keyword{datasets}
