% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{jitterp}
\alias{jitterp}
\title{plot data as a jitter-plot}
\usage{
jitterp(x, y, width, col = "black", alpha = 0.5, ...)
}
\arguments{
\item{x}{x values to plot (if single value and y is a vector, plot is vertical)}

\item{y}{y value at which to plot (if single value, plot is horizontal)}

\item{width}{standard deviation for jitter}

\item{col}{color for points}

\item{alpha}{opacity for points}

\item{...}{other parameters to be passed on to points()}
}
\value{
adds the points to the open plotting device as a jitter plot and returns an invisible list()-object containing the positions of all points
}
\description{
plot data as a jitter-plot
}
\examples{
c(1,2,3,2,3,2,3,4,4)->tmp
hist(tmp)
jitterp(x=tmp, y=1, width=0.1)
}
