% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{rmean}
\alias{rmean}
\title{Calculate a rolling mean for a vector x.}
\usage{
rmean(x, width = 11)
}
\arguments{
\item{x}{Numeric vector for which to calculate the rolling mean.}

\item{width}{Width of the interval over which to calculate rolling mean values. Should be an uneven number (even numbers are coerced into the next-higher uneven number)}
}
\value{
A numeric vector of the same length as x containing the calculated rolling means, with the first and last few values being NA (depending on the setting for width)
}
\description{
Calculate a rolling mean for a vector x.
}
\examples{
rmean(x=c(1,2,3,4,5,6),width=5)
}
