% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_archosauria}
\alias{tree_archosauria}
\title{tree_archosauria}
\format{
An object of class==phylo with 13 tips and 12 internal nodes.
}
\usage{
tree_archosauria
}
\description{
A time-calibrated phylogenetic tree of Archosauria.
}
\keyword{datasets}
