% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{ab.gg}
\alias{ab.gg}
\title{Make a data.frame() that can be used to plot diversity data with density plots, e.g. in ggplot2}
\usage{
ab.gg(data, taxa = NULL, agerange = c(252, 66), precision_ma = 1)
}
\arguments{
\item{data}{list()-object containing occurrence data.frames or single occurrence data.frame()}

\item{taxa}{Selection of taxa to include. If NULL, then abundance is tabulated for each unique factor level of data$tna}

\item{agerange}{Range of geological ages to include in data.frame()}

\item{precision_ma}{Size of intervals (in ma) at which to calculate diversity within the age range.}
}
\value{
A data.frame() with two collumns: ma, for the numerical age, and tax, for the taxon.
}
\description{
Make a data.frame() that can be used to plot diversity data with density plots, e.g. in ggplot2
}
\details{
Each taxon receives one entry per occurrence per time interval. The number of entries per taxon at any given point is thus proportional to the abundance of the taxon in the fossil record, and can be used for plotting with frequency- or density-based functions (e.g. hist(), ggplot2::geom_violin(), etc.). Note that using age values in the original occurrence table instead of this function will usually be more than sufficient and this function thus unnecessary.
}
\examples{
data(archosauria)
ab.gg(archosauria, taxa=c("Pterosauria","Aves"), agerange=c(252,0),precision_ma=1)->flyers
library(ggplot2)
ggplot(data=flyers, aes(x=tax, y=ma))+ylim(252,0)+geom_violin(scale="count")
ggplot(data=flyers, aes(col=tax, x=ma))+xlim(252,0)+geom_density(adjust=0.5)
}
