% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{divdistr_}
\alias{divdistr_}
\title{Calculate total species diversity for any point in time based on a taxon-range table}
\usage{
divdistr_(
  x,
  table = NULL,
  w = rep(1, length(x)),
  smooth = 0,
  max = table$max,
  min = table$min
)
}
\arguments{
\item{x}{A point in time or vector of points in time, in ma, at which species diversity is to be determined.}

\item{table}{A taxon-range table to be used, usually the output of mk.sptab()}

\item{w}{A vector of weights to apply to the estimated (raw) diversity figures. This vector needs to be of the same length as x. Each raw diversity estimate will then be multiplied by the weight. Can be used to account for differences in collection intensity/sampling biases, if these can be quantified (e.g. by analyzing collection records.}

\item{smooth}{The smoothing margin, in units of ma. Corresponds to the plusminus parameter of rmeana(). Defaults to 0, i.e. no smoothing (beyond the resolution determined by the resolution of x)}

\item{max}{Vector or collumn containing the maximum age of each entry in the taxon-range table. Defaults to table$max}

\item{min}{Vector or collumn containing the minimum age of each entry in the taxon-range table. Defaults to table$min}
}
\value{
A numeric vector containing taxon diversity (at the chosen taxonomic level used in the generation of the range table) at the provided ages.
}
\description{
Calculate total species diversity for any point in time based on a taxon-range table
}
\examples{
data(archosauria)
divdistr_(c(170:140),table=archosauria$sptab_Stegosauria)
curve(divdistr_(x,archosauria$sptab_Stegosauria), xlim=c(200,100),ylim=c(-5,35))
ts.stages(ylim=c(-6,-1),alpha=0.3,border=add.alpha("grey"))
ts.periods(ylim=c(-6,-1),alpha=0.0)
}
