% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoDiv.R
\name{ggcol}
\alias{ggcol}
\title{Replicate the standard colour scheme from ggplot2}
\usage{
ggcol(n)
}
\arguments{
\item{n}{Length of colour vector to return.}
}
\value{
A character vector containing color hex codes.
}
\description{
Replicate the standard colour scheme from ggplot2
}
\examples{
ggcol(3)
}
